move to sparkmergetreesink.h/cpp
